🧹 [code health improvement] Refactor suggestionEngine.ts generators
Extracted the massive generation logic from `suggestionEngine.ts` into isolated, focused generator functions in `src/engine/assistant/generators/`. This improves readability and maintainability without altering functionality.
