+++ src/engine/assistant/suggestionEngine.ts
@@ -1,33 +1,18 @@
-import type {
-  ChainLink,
-  EvolutionChain,
-  LocationAreaEncounter,
-  PokemonEncounter,
-  VersionEncounterDetail,
-} from 'pokenode-ts';
+import type { ChainLink, EvolutionChain, LocationAreaEncounter, PokemonEncounter } from 'pokenode-ts';
 import { getGenerationConfig } from '../../utils/generationConfig';
 import { pokeapi } from '../../utils/pokeapi';
-import {
-  GEN1_ITEMS,
-  GEN1_MAP_TO_SLUG,
-  OBEDIENCE_CAPS,
-  STATIC_GIFT_DATA,
-  STATIC_NPC_TRADE_DATA,
-} from '../data/gen1/assistantData';
-import { gen2Items } from '../data/gen2/legacyNameMap';
-import { getUnobtainableReason } from '../exclusives/gen1Exclusives';
-import { getDistanceToMap } from '../mapGraph/gen1Graph';
+import { GEN1_MAP_TO_SLUG, STATIC_GIFT_DATA } from '../data/gen1/assistantData';
 import type { PokemonInstance, SaveData } from '../saveParser/index';
-import type { EncounterDetail, RejectedSuggestion, Suggestion } from './strategies/types';
+import { generateCatchSuggestions } from './generators/catchGenerator';
+import { generateEvolutionSuggestions } from './generators/evolutionGenerator';
+import { generateGiftSuggestions } from './generators/giftGenerator';
+import { generateTradeSuggestions } from './generators/tradeGenerator';
+import type { SuggestionContext } from './generators/types';
+import { generateUtilitySuggestions } from './generators/utilityGenerator';
+import type { RejectedSuggestion, Suggestion } from './strategies/types';
+import { parseIdFromUrl } from './utils';
 
-/** Data returned by fetchAssistantApiData */
-// Construct a reverse map for Gen 2 item names to their IDs.
-// Normalize names for easier matching from PokeAPI (e.g. "thunder-stone" matches "thunderstone")
-const GEN2_ITEM_IDS_BY_NAME: Record<string, number> = {};
-for (const [idStr, name] of Object.entries(gen2Items)) {
-  const normalizedName = name.toLowerCase().replace(/[-\s]/g, '');
-  GEN2_ITEM_IDS_BY_NAME[normalizedName] = parseInt(idStr, 10);
-}
+export { parseIdFromUrl }; // re-exporting in case other modules depend on it via suggestionEngine
 
 export interface AssistantApiData {
   localEncounters: PokemonEncounter[];
@@ -36,19 +21,6 @@ export interface AssistantApiData {
   ancestralEncounters: Record<number, Record<number, LocationAreaEncounter[]>>;
   partyEvolutions: Record<number, EvolutionChain>;
   giftChains: Record<number, EvolutionChain>;
-}
-
-/** Safely extract a numeric ID from a PokéAPI resource URL */
-function parseIdFromUrl(url: string): number {
-  return parseInt(url.split('/').slice(-2, -1)[0] ?? '0', 10);
-}
-
-/**
- * Helper function to find all Pokemon IDs in an evolution chain.
- */
-function getChainIds(node: ChainLink): number[] {
-  const id = parseIdFromUrl(node.species.url);
-  return [id, ...node.evolves_to.flatMap(getChainIds)];
 }
 
 /**
@@ -226,550 +198,26 @@ export function generateSuggestions(
   const displayVersion = effectiveVersion === 'unknown' ? genConfig.defaultVersion : effectiveVersion;
   const queryTargets = missingIds.slice(0, 100);
 
-  const _localSlug = saveData.generation === 1 ? GEN1_MAP_TO_SLUG[saveData.currentMapId] || '' : 'new-bark-town-area';
+  const context: SuggestionContext = {
+    saveData,
+    apiData,
+    missingIds,
+    ownedSet,
+    ownedCount,
+    allInstances,
+    myOtIds,
+    displayVersion,
+    queryTargets,
+    suggestions,
+    rejected,
+    unobtainableCount: 0,
+  };
 
-  // A. Catch logic
-  if (apiData.localEncounters) {
-    const localPids: number[] = [];
-    const localEncounterInfo: Record<number, EncounterDetail[]> = {};
-
-    for (const encounter of apiData.localEncounters) {
-      const pid = parseIdFromUrl(encounter.pokemon.url);
-
-      const isStaticGift = !!STATIC_GIFT_DATA[pid];
-      if (isStaticGift && myOtIds.has(pid)) {
-        rejected.push({
-          pokemonId: pid,
-          reason: 'You already own this one-time gift (matched by Trainer Name).',
-          code: 'GIFT_CLAIMED',
-        });
-        continue;
-      }
-
-      const gift = STATIC_GIFT_DATA[pid];
-      if (saveData.eventFlags && gift?.eventFlag) {
-        if (((saveData.eventFlags[Math.floor(gift.eventFlag / 8)] ?? 0) & (1 << (gift.eventFlag % 8))) !== 0) {
-          rejected.push({
-            pokemonId: pid,
-            reason: `Event flag ${gift.eventFlag} is set. Gift already claimed.`,
-            code: 'GIFT_CLAIMED',
-          });
-          continue;
-        }
-      }
-
-      const versionDetail = encounter.version_details.find(
-        (vd: VersionEncounterDetail) => vd.version.name === displayVersion,
-      );
-      if (versionDetail && missingIds.includes(pid)) {
-        localPids.push(pid);
-        localEncounterInfo[pid] = versionDetail.encounter_details.map((ed) => ({
-          chance: ed.chance,
-          method: ed.method.name,
-          minLevel: ed.min_level,
-          maxLevel: ed.max_level,
-        }));
-      }
-    }
-    if (localPids.length > 0) {
-      const areaName = saveData.currentMapName || 'where you need to be';
-      suggestions.push({
-        id: 'catch-local',
-        category: 'Catch',
-        title: 'Catch Right Here',
-        description: `You are exactly ${areaName}! There are ${localPids.length} missing Pokémon right here.`,
-        pokemonIds: localPids,
-        priority: 120,
-        encounterInfo: localEncounterInfo,
-      });
-    }
-  }
-
-  if (apiData.missingEncounters) {
-    const locationMap: Record<
-      string,
-      {
-        name: string;
-        distance: number;
-        pids: Set<number>;
-        slug: string;
-        encounterMap: Map<number, EncounterDetail[]>;
-      }
-    > = {};
-    let unobtainableCount = 0;
-
-    for (const pid of queryTargets) {
-      const isStaticGift = !!STATIC_GIFT_DATA[pid];
-      if (isStaticGift && myOtIds.has(pid)) continue;
-
-      const gift = STATIC_GIFT_DATA[pid];
-      if (saveData.eventFlags && gift?.eventFlag) {
-        if (((saveData.eventFlags[Math.floor(gift.eventFlag / 8)] ?? 0) & (1 << (gift.eventFlag % 8))) !== 0) continue;
-      }
-
-      const encounters = apiData.missingEncounters[pid] || [];
-      const logicReason = getUnobtainableReason(pid, displayVersion || 'red', ownedCount, ownedSet);
-      if (logicReason) {
-        rejected.push({ pokemonId: pid, reason: logicReason, code: 'CHOICE_TAKEN' });
-        suggestions.push({
-          id: `trade-${pid}`,
-          category: 'Trade',
-          title: `Trade Required: #${pid}`,
-          description: logicReason,
-          pokemonId: pid,
-          priority: 60 - unobtainableCount,
-          debugInfo: {
-            priorityScore: 60 - unobtainableCount,
-            reasoning: 'Explicit unobtainable logic rule',
-          },
-        });
-        unobtainableCount++;
-        continue;
-      }
-
-      let isCatchableSomewhere = encounters.some((enc: LocationAreaEncounter) =>
-        enc.version_details.some((vd: VersionEncounterDetail) => vd.version.name === displayVersion),
-      );
-
-      if (!isCatchableSomewhere) {
-        const aEncsMap = apiData.ancestralEncounters?.[pid] || {};
-        for (const aid in aEncsMap) {
-          if (
-            aEncsMap[Number(aid)]?.some((enc: LocationAreaEncounter) =>
-              enc.version_details.some((vd: VersionEncounterDetail) => vd.version.name === displayVersion),
-            )
-          ) {
-            isCatchableSomewhere = true;
-            break;
-          }
-        }
-
-        if (!isCatchableSomewhere) {
-          const chain = apiData.missingChains?.[pid];
-          const baseId = chain ? parseIdFromUrl(chain.chain.species.url) : pid;
-          const isInternalObtainable = [
-            1,
-            4,
-            7, // Gen 1 Starters
-            152,
-            155,
-            158, // Gen 2 Starters
-            131,
-            133, // Lapras, Eevee
-            138,
-            140,
-            142, // Fossils
-            106,
-            107, // Hitmons
-            143,
-            144,
-            145,
-            146,
-            150, // Gen 1 Legendaries/Snorlax
-            130,
-            185,
-            245,
-            249,
-            250, // Gen 2 Statics
-          ].includes(baseId);
-
-          if (isInternalObtainable) {
-            const isYellow = displayVersion === 'yellow';
-            const isRedBlueStarter = [1, 4, 7].includes(baseId);
-            if (isYellow && isRedBlueStarter) isCatchableSomewhere = true;
-            else if (isInternalObtainable) isCatchableSomewhere = true;
-          }
-        }
-      }
-
-      if (!isCatchableSomewhere && !ownedSet.has(pid)) {
-        // Suppress version exclusive if it can be obtained via NPC trade or Gift
-        const isTradeAvailable = STATIC_NPC_TRADE_DATA.some(
-          (t) =>
-            t.gen === saveData.generation &&
-            t.receivedId === pid &&
-            (!t.versions || t.versions.includes(displayVersion)),
-        );
-
-        const giftEntry = STATIC_GIFT_DATA[pid];
-        const hasDirectGift =
-          !!giftEntry &&
-          (!giftEntry.gen || giftEntry.gen === saveData.generation) &&
-          (!giftEntry.name.includes('Yellow only') || displayVersion === 'yellow');
-
-        if (!isTradeAvailable && !hasDirectGift) {
-          rejected.push({
-            pokemonId: pid,
-            reason: `Not catchable in ${displayVersion} version.`,
-            code: 'VERSION_EXCLUSIVE',
-          });
-          suggestions.push({
-            id: `trade-${pid}`,
-            category: 'Trade',
-            title: `Version Exclusive: #${pid}`,
-            description: `This Pokémon is not available in ${displayVersion}. You must trade for it.`,
-            pokemonId: pid,
-            priority: 50 - unobtainableCount,
-            debugInfo: {
-              priorityScore: 50 - unobtainableCount,
-              reasoning: 'Version exclusivity check',
-            },
-          });
-          unobtainableCount++;
-          continue;
-        }
-      }
-
-      const versionEncounters = encounters.filter((enc: LocationAreaEncounter) =>
-        enc.version_details.some((vd: VersionEncounterDetail) => vd.version.name === displayVersion),
-      );
-
-      for (const enc of versionEncounters) {
-        const targetAreaSlug = enc.location_area.name;
-        if (!locationMap[targetAreaSlug]) {
-          const route = getDistanceToMap(saveData.currentMapId, targetAreaSlug);
-          if (route)
-            locationMap[targetAreaSlug] = {
-              name: route.name,
-              distance: route.distance,
-              pids: new Set(),
-              slug: targetAreaSlug,
-              encounterMap: new Map(),
-            };
-        }
-        if (locationMap[targetAreaSlug]) {
-          locationMap[targetAreaSlug].pids.add(pid);
-          const versionDetail = enc.version_details.find(
-            (vd: VersionEncounterDetail) => vd.version.name === displayVersion,
-          );
-          if (versionDetail) {
-            locationMap[targetAreaSlug].encounterMap.set(
-              pid,
-              versionDetail.encounter_details.map((ed) => ({
-                chance: ed.chance,
-                method: ed.method.name,
-                minLevel: ed.min_level,
-                maxLevel: ed.max_level,
-              })),
-            );
-          }
-        }
-      }
-    }
-    const locations = Object.values(locationMap).map((loc) => ({
-      ...loc,
-      yield: loc.pids.size,
-    }));
-    const partyHasFly = saveData.partyDetails?.some((p: PokemonInstance) => p.moves?.includes(19));
-    locations.sort((a, b) =>
-      partyHasFly
-        ? b.yield !== a.yield
-          ? b.yield - a.yield
-          : a.distance - b.distance
-        : a.distance !== b.distance
-          ? a.distance - b.distance
-          : b.yield - a.yield,
-    );
-    locations.slice(0, 4).forEach((loc) => {
-      const pids = Array.from(loc.pids as Set<number>);
-      const locEncounterInfo: Record<number, EncounterDetail[]> = {};
-
-      pids.forEach((pid) => {
-        const mappedEncounter = loc.encounterMap.get(pid);
-        if (mappedEncounter) {
-          locEncounterInfo[pid] = mappedEncounter;
-        } else {
-          // Check ancestral encounters if not found directly
-          const aEncsMap = apiData.ancestralEncounters?.[pid] || {};
-          for (const aid in aEncsMap) {
-            const ancestorAreaEncounter = aEncsMap[Number(aid)]?.find(
-              (enc: LocationAreaEncounter) =>
-                enc.location_area.name === loc.slug &&
-                enc.version_details.some((vd: VersionEncounterDetail) => vd.version.name === displayVersion),
-            );
-            if (ancestorAreaEncounter) {
-              const versionDetail = ancestorAreaEncounter.version_details.find(
-                (vd: VersionEncounterDetail) => vd.version.name === displayVersion,
-              );
-              if (versionDetail) {
-                locEncounterInfo[pid] = versionDetail.encounter_details.map((ed) => ({
-                  chance: ed.chance,
-                  method: ed.method.name,
-                  minLevel: ed.min_level,
-                  maxLevel: ed.max_level,
-                }));
-                break;
-              }
-            }
-          }
-        }
-      });
-
-      if (loc.distance === 0) {
-        // This is already handled by the localEncounters block at the top with higher priority
-        return;
-      }
-      suggestions.push({
-        id: `catch-loc-${loc.name}`,
-        category: 'Catch',
-        title: `Travel to ${loc.name}`,
-        description: partyHasFly
-          ? `Use Fly to easily travel to ${loc.name} and catch ${pids.length} missing Pokémon!`
-          : `Catch ${pids.length} missing Pokémon at ${loc.name}, only ${loc.distance === 1 ? '1 connection away' : `${loc.distance} connections away`}.`,
-        pokemonIds: pids,
-        priority: 80 + loc.yield - (partyHasFly ? 0 : loc.distance),
-        encounterInfo: locEncounterInfo,
-      });
-    });
-  }
-
-  // NPC Trades
-  for (const trade of STATIC_NPC_TRADE_DATA) {
-    if (trade.gen !== saveData.generation) continue;
-    if (trade.versions && !trade.versions.includes(displayVersion)) continue;
-
-    let isClaimed = false;
-
-    // Gen 1 specific bit flag check
-    if (saveData.generation === 1 && trade.tradeIndex !== undefined && saveData.npcTradeFlags !== undefined) {
-      isClaimed = (saveData.npcTradeFlags & (1 << trade.tradeIndex)) !== 0;
-    } else {
-      // Fallback for Gen 2 or if bit flags are missing
-      isClaimed = allInstances.some(
-        (p: PokemonInstance) => p.speciesId === trade.receivedId && p.otName === trade.receivedOtName,
-      );
-    }
-
-    if (!isClaimed && missingIds.includes(trade.receivedId)) {
-      const hasOffered = ownedSet.has(trade.offeredId);
-      if (hasOffered) {
-        suggestions.push({
-          id: `npc-trade-${trade.receivedId}`,
-          category: 'Trade',
-          title: `Trade for #${trade.receivedId}`,
-          description: `You have #${trade.offeredId}! Trade it at ${trade.location} for #${trade.receivedId}.`,
-          pokemonId: trade.receivedId,
-          priority: 85,
-        });
-      } else {
-        suggestions.push({
-          id: `npc-trade-${trade.receivedId}`,
-          category: 'Trade',
-          title: `Trade for #${trade.receivedId}`,
-          description: `Catch #${trade.offeredId} and trade it at ${trade.location} for #${trade.receivedId}.`,
-          pokemonId: trade.receivedId,
-          priority: 65,
-        });
-      }
-    }
-  }
-
-  // Gifts/Statics
-  for (const [pidStr, gift] of Object.entries(STATIC_GIFT_DATA)) {
-    const pid = parseInt(pidStr, 10);
-    if (gift.gen && gift.gen !== saveData.generation) continue;
-    if (gift.name.includes('Yellow only') && displayVersion !== 'yellow') continue;
-    if (gift.reason.includes('Crystal') && displayVersion !== 'crystal') continue;
-
-    const giftChain = apiData.giftChains?.[pid];
-    const familyIds = giftChain ? getChainIds(giftChain.chain).filter((id) => id <= genConfig.maxDex) : [pid];
-
-    const hasAnyFamily = familyIds.some((fid) => ownedSet.has(fid));
-    const hasAnyWithMyOT = familyIds.some((fid) => myOtIds.has(fid));
-
-    let isClaimedByEvent = false;
-    if (saveData.eventFlags && gift.eventFlag) {
-      isClaimedByEvent =
-        ((saveData.eventFlags[Math.floor(gift.eventFlag / 8)] ?? 0) & (1 << (gift.eventFlag % 8))) !== 0;
-    }
-
-    if (!isClaimedByEvent && !hasAnyWithMyOT && !hasAnyFamily && missingIds.includes(pid)) {
-      suggestions.push({
-        id: `gift-${pid}`,
-        category: 'Gift',
-        title: `Secure Gift: ${gift.name}`,
-        description: `Location: ${gift.location}. ${gift.reason}`,
-        pokemonId: pid,
-        priority: 85,
-      });
-    }
-  }
-
-  // Box Full
-  if (saveData.generation === 1) {
-    if (saveData.currentBoxCount >= 20) {
-      suggestions.push({
-        id: 'utility-box-full',
-        category: 'Utility',
-        title: 'CRITICAL: PC Box Full!',
-        description: "Your current PC box is at 20/20. Switch boxes via Bill's PC.",
-        priority: 150,
-      });
-    } else if (saveData.currentBoxCount >= 18) {
-      suggestions.push({
-        id: 'utility-box-near-full',
-        category: 'Utility',
-        title: 'PC Box Almost Full',
-        description: `${20 - saveData.currentBoxCount} slots remaining.`,
-        priority: 95,
-      });
-    }
-  }
-
-  // Obedience
-  const totalBadges =
-    saveData.generation === 1 ? saveData.badges : (saveData.johtoBadges || 0) + (saveData.kantoBadges || 0);
-  const caps = OBEDIENCE_CAPS.filter((c) => totalBadges >= c.badges);
-  const currentCap = caps[caps.length - 1]?.level ?? 10;
-  const disobedient = saveData.partyDetails.filter(
-    (p) => p.otName && p.otName !== saveData.trainerName && p.level > currentCap,
-  );
-  if (disobedient.length > 0) {
-    suggestions.push({
-      id: 'utility-obedience-danger',
-      category: 'Utility',
-      title: 'Obedience Danger!',
-      description: `You have traded Pokémon above Lv. ${currentCap}. They may not obey you!`,
-      priority: 110,
-    });
-  }
-
-  // Pre-calculate instances by species for O(1) lookup during evolution checks
-  // ⚡ Bolt: Replaces O(N*M) filtering inside the loop below
-  const instancesBySpecies = new Map<number, PokemonInstance[]>();
-  for (const p of allInstances) {
-    if (!instancesBySpecies.has(p.speciesId)) {
-      instancesBySpecies.set(p.speciesId, []);
-    }
-    instancesBySpecies.get(p.speciesId)!.push(p);
-  }
-
-  // Evolutions
-  queryTargets.forEach((targetId: number) => {
-    const chain = apiData.missingChains?.[targetId];
-    if (!chain) return;
-
-    // Find the target node and its parent (pre-evolution)
-    const findNodeAndParent = (
-      node: ChainLink,
-      parent: ChainLink | null = null,
-    ): { targetNode: ChainLink; parentNode: ChainLink } | null => {
-      const id = parseIdFromUrl(node.species.url);
-      // biome-ignore lint/style/noNonNullAssertion: Guaranteed by recursive logic structure
-      if (id === targetId) return { targetNode: node, parentNode: parent! };
-      for (const child of node.evolves_to) {
-        const res = findNodeAndParent(child, node);
-        if (res) return res;
-      }
-      return null;
-    };
-
-    const nodes = findNodeAndParent(chain.chain);
-    if (!nodes?.parentNode) return;
-
-    const parentId = parseIdFromUrl(nodes.parentNode.species.url);
-
-    // Check if we own the pre-evolution
-    // ⚡ Bolt: Use pre-calculated map instead of array filter
-    const ownedInstances = instancesBySpecies.get(parentId) || [];
-    if (ownedInstances.length === 0) return;
-
-    // Get the highest level instance to give the most optimistic suggestion
-    const bestInstance = ownedInstances.reduce((prev, current) => (prev.level > current.level ? prev : current));
-
-    // Get evolution details from the target node
-    const details = nodes.targetNode.evolution_details[0];
-    if (!details) return;
-
-    const trigger = details.trigger.name;
-
-    if (trigger === 'level-up') {
-      if (details.min_level) {
-        const isReady = bestInstance.level >= details.min_level;
-        let statCondition = '';
-        if (details.relative_physical_stats === 1) statCondition = ' (needs Attack > Defense)';
-        else if (details.relative_physical_stats === -1) statCondition = ' (needs Attack < Defense)';
-        else if (details.relative_physical_stats === 0) statCondition = ' (needs Attack = Defense)';
-
-        suggestions.push({
-          id: `evo-lvl-${targetId}`,
-          category: 'Evolve',
-          title: `Level Up Evolution: #${targetId}`,
-          description: isReady
-            ? `Your Lv. ${bestInstance.level} pre-evolution is ready to evolve (needs Lv. ${details.min_level})${statCondition}!`
-            : `Your Lv. ${bestInstance.level} pre-evolution evolves at Lv. ${details.min_level}${statCondition}.`,
-          pokemonId: targetId,
-          priority: isReady ? 90 : 75,
-        });
-      } else if (details.min_happiness) {
-        let timeCondition = '';
-        if (details.time_of_day === 'Day' || (details.time_of_day as unknown as string) === 'day')
-          timeCondition = ' during the day';
-        else if (details.time_of_day === 'Night' || (details.time_of_day as unknown as string) === 'night')
-          timeCondition = ' during the night';
-
-        suggestions.push({
-          id: `evo-happy-${targetId}`,
-          category: 'Evolve',
-          title: `Happiness Evolution: #${targetId}`,
-          description: `Level up your pre-evolution with high happiness${timeCondition} to evolve!`,
-          pokemonId: targetId,
-          priority: 80,
-        });
-      }
-    } else if (trigger === 'use-item' && details.item) {
-      const itemName = details.item.name;
-      const isYellowStarterPikachu =
-        displayVersion === 'yellow' && parentId === 25 && bestInstance.otName === saveData.trainerName;
-      if (isYellowStarterPikachu) return;
-
-      let targetItemId = -1;
-      if (saveData.generation === 1) {
-        if (itemName.includes('fire')) targetItemId = GEN1_ITEMS.FIRE_STONE;
-        else if (itemName.includes('thunder')) targetItemId = GEN1_ITEMS.THUNDER_STONE;
-        else if (itemName.includes('water')) targetItemId = GEN1_ITEMS.WATER_STONE;
-        else if (itemName.includes('leaf')) targetItemId = GEN1_ITEMS.LEAF_STONE;
-        else targetItemId = GEN1_ITEMS.MOON_STONE; // Default to moon stone
-      } else {
-        // Gen 2
-        const normalizedItemName = itemName.toLowerCase().replace(/[-\s]/g, '');
-        targetItemId = GEN2_ITEM_IDS_BY_NAME[normalizedItemName] || -1;
-      }
-
-      const hasStone = saveData.inventory.some((i) => i.id === targetItemId);
-      if (hasStone) {
-        suggestions.push({
-          id: `evo-stn-${targetId}`,
-          category: 'Evolve',
-          title: `Ready to Evolve: #${targetId}!`,
-          description: `Use ${itemName.replace('-', ' ')} on your pre-evolution!`,
-          pokemonId: targetId,
-          priority: 95,
-        });
-      } else {
-        suggestions.push({
-          id: `evo-buy-${targetId}`,
-          category: 'Evolve',
-          title: `Get ${itemName.replace('-', ' ')}`,
-          description: `Obtain a ${itemName.replace('-', ' ')} to evolve your pre-evolution into #${targetId}.`,
-          pokemonId: targetId,
-          priority: 40,
-        });
-      }
-    } else if (trigger === 'trade') {
-      const heldItem = details.held_item?.name;
-      suggestions.push({
-        id: `evo-trade-${targetId}`,
-        category: 'Evolve',
-        title: `Trade Evolution: #${targetId}`,
-        description: heldItem
-          ? `Trade your pre-evolution while holding ${heldItem.replace('-', ' ')}.`
-          : `Trade your pre-evolution to evolve it!`,
-        pokemonId: targetId,
-        priority: 85,
-      });
-    }
-  });
+  generateCatchSuggestions(context);
+  generateTradeSuggestions(context);
+  generateGiftSuggestions(context);
+  generateUtilitySuggestions(context);
+  generateEvolutionSuggestions(context);
 
   const uniqueSuggestions = Array.from(new Map(suggestions.map((item) => [item.id, item])).values());
   uniqueSuggestions.sort((a, b) => b.priority - a.priority);
